requirements.txt
pymupdf>=1.24.0
ezdxf>=0.18.0
cryptography
ntplib
Pillow
pyinstaller
